cupy module cleanup
