+++ wx_cupy/linalg.py
